+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-07T10:50:00.302Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-07T12:12:06.586Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")