+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9EZ0AZKGA8SX5E397738E9K"
+        "01K9F3PP1NBW5KGBMSZQY9DGVQ"
     )
     r"""The unique identifier of the data source"""