+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-07T10:50:00.302Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-07T12:12:06.586Z")
     r"""The date and time the resource was last updated"""