+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9EZ0AZGQG8HE0E468V8BEHF"
+        "01K9F3PP1JPGJB207BD0PY4E44"
     )
     r"""The unique identifier of the data source"""