+++ packages/orq-rc/RELEASES.md
@@ -3899,3 +3899,13 @@ Based on:
 - [python v4.0.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.15 - packages/orq-rc
+
+## 2025-11-07 12:14:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.650.0 (2.743.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.16 - packages/orq-rc